### 2025-03-31 #7962 @editorialbot
[PRE REVIEW]: WunDeeDB.jl: A easy to use, zero config, WAL, SQLite backend vector database
**Submitting author:** @mantzaris (<a href="http://orcid.org/0000-0002-0026-5725">Alexander V. Mantzaris</a>)
**Repository:** https://github.com/mantzaris/WunDeeDB.jl
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
