### 2025-04-02 #7962 @mantzaris
@jbytecode I will get those paper edits done soon. In regards to the commit history, I worked on it for some time experimenting on different approaches for the on disk memory management. I put the code into this repository once I had settled on an approach I had pretty much already explored. It did not seem like those commits were going to be useful.
