+++ paper/paper.md
@@ -20,7 +20,7 @@ bibliography: paper.bib
 
 # Summary
 
-WunDeeDB.jl is a Julia package that provides a disk-backed system for storing, searching, and managing embedding vectors at scale, influenced by disk-oriented graph-based ANN techniques [@pan2023lm; @jayaram2019diskann; @singh2021freshdiskann] and the broader insights from hierarchical small-world graphs [@malkov2018efficient; @wang2021comprehensive]. By maintaining embeddings in a SQLite database and optionally using graph-based indices, WunDeeDB.jl minimizes in-memory overhead while supporting efficient similarity searches on commodity hardware. Its design also facilitates integration with common vector-database or ML pipelines that rely on embedding retrieval.
+WunDeeDB.jl is a package written in Julia Lang that provides a disk-backed system for storing, searching, and managing embedding vectors at scale, influenced by disk-oriented graph-based ANN techniques [@pan2023lm; @jayaram2019diskann; @singh2021freshdiskann] and the broader insights from hierarchical small-world graphs [@malkov2018efficient; @wang2021comprehensive]. By maintaining embeddings in a SQLite database and optionally using graph-based indices, WunDeeDB.jl minimizes in-memory overhead while supporting efficient similarity searches on commodity hardware. Its design also facilitates integration with common vector-database or ML pipelines that rely on embedding retrieval.
 
 In contrast to fully in-memory approaches, WunDeeDB.jl leverages disk-based storage and user-configurable adjacency (e.g., HNSW, LM-DiskANN, or fallback linear search), allowing large-scale data to be handled without saturating RAM. It supports incremental insertions and deletions, ensuring the index remains up-to-date as datasets evolve. By combining these disk-native strategies with tunable BFS expansions and adjacency pruning, WunDeeDB.jl enables robust nearest neighbor searches for high-dimensional embeddings.
 
@@ -31,7 +31,7 @@ Features include:
 - **Incremental insert/delete** operations that update disk-based adjacency structures to keep pace with dataset changes.
 - **Configurable BFS expansions** (e.g., `EF_SEARCH`, `EF_CONSTRUCTION`) to balance search speed vs. recall.
 
-With these capabilities, **WunDeeDB.jl** offers a practical and scalable solution for disk-based embedding management, building on research showing the viability of disk-native approaches for large ANN indexes [@pan2023lm; @jayaram2019diskann; @singh2021freshdiskann]. 
+With these capabilities, **WunDeeDB.jl** offers a practical and scalable solution for disk-based embedding management, building on research showing the viability of disk-native approaches for large ANN indexes [@pan2023lm; @jayaram2019diskann; @singh2021freshdiskann]. It is designed for minimal configuration, providing Write-Ahead Logging (WAL) and transactions. 
 
 
 # Statement of Need
@@ -43,4 +43,11 @@ Approximate Nearest Neighbor (ANN) search is a key element in recommendation sys
 3. **High Recall**: Adjusting BFS expansions and adjacency parameters yields near state-of-the-art accuracy in neighbor retrieval.  
 4. **Scalable & Simple Architecture**: Built using Julia’s performance ecosystem, WunDeeDB.jl integrates disk operations with numeric libraries, and is straightforward to install.
 
-This approach benefits practitioners needing large-scale nearest neighbor indices without requiring specialized clusters or massive RAM. WunDeeDB.jl’s minimal setup, clear usage patterns, and example-driven documentation lower the barrier to entry for embedding-based ANN workflows in resource-constrained environments.
+This approach benefits practitioners needing large-scale nearest neighbor indices without requiring specialized clusters or massive RAM. 
+
+
+# Acknowledgements
+
+Thanks to the Julia community for their continued support of open-source scientific computing.
+
+# References
