+++ paper/paper.bib
@@ -22,7 +22,7 @@
   journal={Advances in neural information processing Systems},
   volume={32},
   year={2019},
-  doi={https://doi.org/10.1145/3543507.3583552}
+  doi={10.1145/3543507.3583552}
 }
 
 @article{wang2021comprehensive,
@@ -30,7 +30,7 @@
   author={Wang, Mengzhao and Xu, Xiaoliang and Yue, Qiang and Wang, Yuxiang},
   journal={arXiv preprint arXiv:2101.12631},
   year={2021},
-  doi={https://doi.org/10.14778/3476249.3476255}
+  doi={10.14778/3476249.3476255}
 }
 
 @article{malkov2018efficient,
