### 2025-04-02 #7962 @mantzaris
@jbytecode , the reference section is there now, and I think the doi urls are fixed to the way they should be [paper](https://github.com/mantzaris/WunDeeDB.jl/blob/main/paper/paper.pdf).

For reviewers I recommend: @ jromanowska , @ MichaelHopwood, @ sglyon, @ tlaceby, @ jjasser87
