### 2025-04-06 #7962 @jbytecode
👋👋👋 Dear @tlaceby  👋👋👋

**I need one more reviewer for this submission.**

**[WunDeeDB.jl: A easy to use, zero config, WAL, SQLite backend vector database](https://github.com/openjournals/joss-reviews/issues/7962)**

You can find more information at the top of this Github issue ([LINK](https://github.com/openjournals/joss-reviews/issues/7962)).

Thank you in advance!
