### 2025-04-20 c0a8aff
paper improvement by including more citations and reinforcing the need statement
