+++ paper/paper.bib
@@ -25,6 +25,14 @@
   doi={10.1145/3543507.3583552}
 }
 
+@misc{diskann-github,
+   author = {Simhadri, Harsha Vardhan and Krishnaswamy, Ravishankar and Srinivasa, Gopal and Subramanya, Suhas Jayaram and Antonijevic, Andrija and Pryce, Dax and Kaczynski, David and Williams, Shane and Gollapudi, Siddarth and Sivashankar, Varun and Karia, Neel and Singh, Aditi and Jaiswal, Shikhar and Mahapatro, Neelam and Adams, Philip and Tower, Bryan and Patel, Yash},
+   title = {DiskANN: Graph-structured Indices for Scalable, Fast, Fresh and Filtered Approximate Nearest Neighbor Search},
+   url = {https://github.com/Microsoft/DiskANN},
+   version = {0.6.1},
+   year = {2023}
+}
+
 @article{wang2021comprehensive,
   title={A comprehensive survey and experimental comparison of graph-based approximate nearest neighbor search},
   author={Wang, Mengzhao and Xu, Xiaoliang and Yue, Qiang and Wang, Yuxiang},
@@ -43,4 +51,14 @@
   year={2018},
   publisher={IEEE},
   doi={10.1109/TPAMI.2018.2889473}
+}
+
+@article{tellez2022similaritysearch,
+  title={Similaritysearch. jl: autotuned nearest neighbor indexes for Julia},
+  author={Tellez, Eric S and Ruiz, Guillermo},
+  journal={Journal of Open Source Software},
+  volume={7},
+  number={75},
+  pages={4442},
+  year={2022}
 }
