### 2025-05-03 #8033 @tlaceby
Yes, I've already shared some feedback with the author, and the suggested changes have been made to the repository.

Here's one additional suggestion that I believe would improve the readability of the documentation:

1. **Include a link to the documentation** in the **README.md** where the documentation is referenced. For example, just above [[naming of the package](https://github.com/mantzaris/WunDeeDB.jl?tab=readme-ov-file#naming-of-the-package)](https://github.com/mantzaris/WunDeeDB.jl?tab=readme-ov-file#naming-of-the-package), it might be helpful to include a direct link to the full documentation so users can access it more easily.
