### 2025-05-04 #8033 @jbytecode
@sglyon - Would you like me to give an extra time for your review (yes, the review has just started and we have time to finalize), or, I can remove you reviewer identity and find someone else? It would be nice to see your contribution - JOSS goes alive with your contributions. Thank you in advance.
