### 2025-05-22 #8033 @jbytecode
Thank you, please create your review list and go on.

You can start with typing

`@editorialbot generate my checklist`
