### 2025-06-15 3fe3ef7
minor fixes for the manuscript and the bibliography
