+++ paper/paper.bib
@@ -54,11 +54,26 @@
 }
 
 @article{tellez2022similaritysearch,
-  title={Similaritysearch. jl: autotuned nearest neighbor indexes for Julia},
+  title={Similaritysearch.jl: {A\utotuned nearest neighbor indexes for {J}ulia},
   author={Tellez, Eric S and Ruiz, Guillermo},
   journal={Journal of Open Source Software},
   volume={7},
   number={75},
   pages={4442},
-  year={2022}
+  year={2022},
+  doi={10.21105/joss.04442}
+}
+
+@article{julia,
+  doi       = {10.1137/141000671},
+  url       = {https://doi.org/10.1137%2F141000671},
+  year      = 2017,
+  month     = {jan},
+  publisher = {Society for Industrial {\&} Applied Mathematics ({SIAM})},
+  volume    = {59},
+  number    = {1},
+  pages     = {65--98},
+  author    = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
+  title     = {{J}ulia: A Fresh Approach to Numerical Computing},
+  journal   = {{SIAM} Review}
 }
