### 2025-06-15 d7812a9
fix bibtex
