+++ paper/paper.bib
@@ -54,7 +54,7 @@
 }
 
 @article{tellez2022similaritysearch,
-  title={Similaritysearch.jl: {A\utotuned nearest neighbor indexes for {J}ulia},
+  title={Similaritysearch.jl: {A}utotuned nearest neighbor indexes for {J}ulia},
   author={Tellez, Eric S and Ruiz, Guillermo},
   journal={Journal of Open Source Software},
   volume={7},
