### 2025-06-15 #8033 @mantzaris
@jbytecode , I agree with them and the pull request has been merged.
