### 2025-06-16 #8033 @jbytecode
@mantzaris - The paper and the bibtex look better, thank you for merging.

Now we need a zenodo archive of the reviewed package. Please create a tagged release in the software repository, e.g. v1.2.3. Use the release id to create a Zenodo archive of the reviewed software. The metadata including license, title, author list, ORCID should match those in the manuscript and software.

Please report me the archive link, the DOI, and the release.

Thank you in advance.
