+++ v1.0.0/search_index.js
@@ -1,3 +1,3 @@
 var documenterSearchIndex = {"docs":
-[{"location":"","page":"Home","title":"Home","text":"CurrentModule = WunDeeDB","category":"page"},{"location":"","page":"Home","title":"Home","text":"","category":"page"},{"location":"#WunDeeDB","page":"Home","title":"WunDeeDB","text":"","category":"section"},{"location":"","page":"Home","title":"Home","text":"WunDeeDB is a Julia package for storing and querying embeddings in a SQLite database. It supports a variety of numerical types (including Float16, Float32, Float64, and various integer types) and can integrate with approximate nearest neighbor indices (like HNSW or LM‐DiskANN). By default, the module provides a linear fallback for distance queries if no ANN method is enabled.","category":"page"},{"location":"","page":"Home","title":"Home","text":"This module exposes bulk operations for insertions, deletions, and updates (capped at a certain transaction size) and provides convenient high‐level functions (insert_embeddings, delete_embeddings, search_ann, etc.) so users can focus on storing and retrieving embeddings without worrying about the underlying details of adjacency structures. Whether you want a simple linear search or a more scalable approach with HNSW or LM‐DiskANN, WunDeeDB automatically handles the index creation and updates.","category":"page"},{"location":"","page":"Home","title":"Home","text":"Internally, WunDeeDB stores embedding vectors in a BLOB column, along with metadata (embedding_length, data_type) in a meta table to ensure consistency. When you choose ann=\"hnsw\" or ann=\"lmdiskann\", each call to insert_embeddings automatically triggers adjacency insertion for that node, enabling fast approximate searches for new queries via search_ann.","category":"page"},{"location":"","page":"Home","title":"Home","text":"","category":"page"},{"location":"#Supported-Data-Types","page":"Home","title":"Supported Data Types","text":"","category":"section"},{"location":"","page":"Home","title":"Home","text":"WunDeeDB supports storing embeddings of these types (any dimension in principle):","category":"page"},{"location":"","page":"Home","title":"Home","text":"Float16, Float32, Float64, BigFloat, Int8, UInt8, Int16, UInt16, Int32, UInt32, Int64, UInt64, Int128, UInt128","category":"page"},{"location":"","page":"Home","title":"Home","text":"","category":"page"},{"location":"#Exported-Functions","page":"Home","title":"Exported Functions","text":"","category":"section"},{"location":"","page":"Home","title":"Home","text":"General DB  ","category":"page"},{"location":"","page":"Home","title":"Home","text":"initialize_db, open_db, close_db, delete_db, delete_all_embeddings  \nget_meta_data, update_description, count_entries","category":"page"},{"location":"","page":"Home","title":"Home","text":"Embeddings  ","category":"page"},{"location":"","page":"Home","title":"Home","text":"insert_embeddings, delete_embeddings, update_embeddings  \nget_embeddings, get_all_ids, get_all_embeddings, random_embeddings","category":"page"},{"location":"","page":"Home","title":"Home","text":"ANN / Searching  ","category":"page"},{"location":"","page":"Home","title":"Home","text":"search_ann (resolves to HNSW, LM‐DiskANN, or linear fallback)  \nsupported_distance_metrics","category":"page"},{"location":"","page":"Home","title":"Home","text":"Linear Search  ","category":"page"},{"location":"","page":"Home","title":"Home","text":"linear_search_all_embeddings, linear_s
[…]
