### 2025-06-17 #8033 @jbytecode
@mantzaris - I changed the title of this review issue. (I think `An easy` is the correct form of `A easy`).

Are you agree with this?

If so, could you please change the title of the manuscript and the archive?

Thank you in advance.
