### 2025-06-17 #8033 @mantzaris
@jbytecode , Ok, I think I got it to save the update now, and that the change is visible on your side too:

- https://figshare.com/articles/dataset/WunDeeDB_jl_A_easy_to_use_zero_config_WAL_SQLite_backend_vector_database/29330072?file=55416563
- 10.6084/m9.figshare.29330072
