### 2025-06-17 #8033 @jbytecode
@mantzaris - okay, now I'm recommending an acceptance. Our track editor will make the final decision. Please hold on.

Thank you for submitting your work to JOSS!
