### 2025-08-13 #7962 @jjasser87
> 👋👋👋 Dear [@jjasser87](https://github.com/jjasser87) 👋👋👋
>
> **I need one more reviewer for this submission.**
>
> Would you be willing to assist in reviewing this submission for **JOSS** (**Journal of Open Source Software**)?
>
> JOSS publishes articles about open source research software. The submission I'd like you to review is titled:
>
> **[WunDeeDB.jl: A easy to use, zero config, WAL, SQLite backend vector database](https://github.com/openjournals/joss-reviews/issues/7962)**
>
> You can find more information at the top of this Github issue ([LINK](https://github.com/openjournals/joss-reviews/issues/7962)).
>
> The review process at JOSS is unique: it takes place in a GitHub issue, is open, and author-reviewer-editor conversations are encouraged. If you have any questions please let me know.
>
> This is the pre-review issue. After setting at least 2 reviewers we will start the review process in a separate thread. In that thread, there will be 23 check items for each single reviewer.
>
> Thank you in advance!

So Sorry for the way too late reply. Just noticed this message. However, I wouldn’t have been able to review the paper anyway, as I know the author and have previously worked with them.
